fix: JWT 토큰 설정 및 영상 리스트 조회 기능 수정

+++ meerkatai/src/main/java/com/capstone/meerkatai/video/dto/VideoDetailsResponse.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class VideoDetailsResponse {
-    private Integer video_id;
+    private Long video_id;
     private String file_path;
     private String thumbnail_path;
     private Integer duration;
