+++ meerkatai/src/main/java/com/capstone/meerkatai/video/entity/Video.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class Video {
 
     @Id
-    private Integer videoId;
+    private Long videoId;
 
     @Column(nullable = false, length = 250)
     private String filePath;
